requirements.txt
# Core Dependencies
pyspark>=3.3.0
delta-spark>=2.3.0
databricks-connect>=13.0

# Data Processing
pandas>=1.5.0
numpy>=1.23.0

# Email and Communication
smtplib-status>=1.0.3
email-validator>=2.0.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0

# Development Tools
black>=22.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Monitoring and Logging
logging>=0.4.9.6
python-json-logger>=2.0.7 